Revert "Update to v1.1"
