+++ KlebLib/polynomial.py
@@ -1,19 +1,28 @@
-"""Polynomial -- store and manipulate polynomials"""
-
 import re
+from typing import Union
 from copy import deepcopy
 
 __all__ = ['Polynomial']
 
-class Polynomial:
-    """Store a polynomial as a list of terms.
+def get_variables(polynomial:Union[list, str]) -> list:
+    variables = set()
 
-    Methods:
-    differentiate(varToDiff) -- differentiates the polynomial with respect to varToDiff. if no variable is supplied, and the polynomial only has a single variable, that will be used instead
-    integrate(varToIntegrate) -- integrates the polynomial with respect to varToIntegrate if no variable is supplied, and the polynomial only has a single variable, that will be used instead
-    integrate_definite(varToIntegrate)
-    """
-    def __init__(self, polynomial:str|list):
+    if type(polynomial) is str:
+        for character in polynomial:
+            #If it is a letter
+            if (ord(character) >= 65 and ord(character) <= 90) or (ord(character) >= 97 and ord(character) <= 122):
+                variables.add(character)
+
+    elif type(polynomial) is list:
+        for term in polynomial:
+            for variable in term:
+                if variable != 'num':
+                    variables.add(variable)
+
+    return list(variables)
+
+class Polynomial:
+    def __init__(self, polynomial:Union[list, str]):
         #print(f'the polynomials are {polynomial} and are of type {type(polynomial).__name__}') #debug
         if type(polynomial) is list:
             for term in polynomial:
@@ -21,10 +30,10 @@ class Polynomial:
                     raise KeyError('Every term must contain key \'num\'')
             self.polynomial = polynomial
         else:
-            self.polynomial = self._parse(polynomial, Polynomial.get_variables(polynomial))
+            self.polynomial = self._parse(polynomial, get_variables(polynomial))
         #print(self.polynomial) #debug
 
-    def _parse(self, polynomial:str) -> list:
+    def _parse(self, polynomial:str, variables:list) -> list:
         #parse the polynomial into a list of dictionaries
         
         #print(f'parsing polynomial {polynomial}') #debug
@@ -59,7 +68,7 @@ class Polynomial:
             else:
                 negativeMultiple = 1
                         
-            for variable in Polynomial.get_variables(term):
+            for variable in get_variables(term):
                 if f'{variable}^' in term:
                     currentTerm[variable] = self._trim_num(term, 'right')
                             
@@ -106,10 +115,10 @@ class Polynomial:
 
     def integrate(self, varToIntegrate:str=None):
         if varToIntegrate is None:
-            if len(Polynomial.get_variables(self.polynomial)) != 1:
+            if len(get_variables(self.polynomial)) != 1:
                 raise TypeError('cannot implicitly detect variable for polynomials of multiple variables')
             else:
-                varToIntegrate = Polynomial.get_variables(self.polynomial)[0]
+                varToIntegrate = get_variables(self.polynomial)[0]
                 
         outputPolynomial = []
         for term in self.polynomial:
@@ -152,7 +161,7 @@ class Polynomial:
                 output.append(outputTerm)
 
             #print(f'output {i} before cleanup is {output}') #debug
-            output = Polynomial.consolidate(output)
+            output = self._consolidate(output)
             #print(f'output {i} after cleanup is {output}') #debug
 
         upper = Polynomial(outputs[0])
@@ -166,14 +175,14 @@ class Polynomial:
         
         for i, term in enumerate(deepcopy(other.polynomial)):
             #print(f'adding term {term} to polynomial {outputPolynomial}') #debug
-            location = Polynomial.locate(outputPolynomial.copy(), term.copy())
+            location = self._locate(outputPolynomial.copy(), term.copy())
             if not type(location) is bool:
                 outputPolynomial[location]['num'] += term['num']
             else:
                 outputPolynomial.append(term.copy())
 
         #print(f'finished polynomial is {self.consolidate(outputPolynomial)}') #debug
-        return Polynomial(Polynomial.consolidate(outputPolynomial))
+        return Polynomial(self._consolidate(outputPolynomial))
 
     def __sub__(self, other):
         #print(f'subtracting polynomial {other} from {self}') #debug
@@ -181,7 +190,7 @@ class Polynomial:
         
         for i, term in enumerate(deepcopy(other.polynomial)):
             #print(f'subtracting term {term} from polynomial {outputPolynomial}') #debug
-            location = Polynomial.locate(outputPolynomial.copy(), term.copy())
+            location = self._locate(outputPolynomial.copy(), term.copy())
             if not type(location) is bool:
                 outputPolynomial[location]['num'] -= term['num']
             else:
@@ -194,7 +203,7 @@ class Polynomial:
                 outputPolynomial.append(currentTerm)
 
         #print(f'finished polynomial is {self.consolidate(outputPolynomial)}') #debug
-        return Polynomial(Polynomial.consolidate(outputPolynomial))
+        return Polynomial(self.consolidate(outputPolynomial))
 
     def __iadd__(self, other):
         #print(f'adding polynomials {self} and {other}') #debug
@@ -202,7 +211,7 @@ class Polynomial:
         
         for i, term in enumerate(deepcopy(other.polynomial)):
             #print(f'adding term {term} to polynomial {outputPolynomial}') #debug
-            location = Polynomial.locate(outputPolynomial.copy(), term.copy())
+            location = self._locate(outputPolynomial.copy(), term.copy())
             if not type(location) is bool:
                 outputPolynomial[location]['num'] += term['num']
             else:
@@ -218,8 +227,8 @@ class Polynomial:
         
         for i, term in enumerate(deepcopy(other.polynomial)):
             #print(f'suntracting term {term} from polynomial {outputPolynomial}') #debug
-            location = Polynomial.locate(outputPolynomial.copy(), term.copy())
-            if location is not None:
+            location = self._locate(outputPolynomial.copy(), term.copy())
+            if not type(location) is bool:
                 outputPolynomial[location]['num'] -= term['num']
             else:
                 currentTerm = {}
@@ -231,7 +240,7 @@ class Polynomial:
                 outputPolynomial.append(currentTerm)
 
         #print(f'finished polynomial is {self.consolidate(outputPolynomial)}') #debug
-        self.polynomial = Polynomial.consolidate(outputPolynomial)
+        self.polynomial = self.consolidate(outputPolynomial)
         return self
 
     def __str__(self):
@@ -242,9 +251,9 @@ class Polynomial:
             if term['num'] < 0:
                 if term['num'] != -1:
                     if i == 0:
-                        output += f'-{Polynomial.int_if_pos(abs(term["num"]))}'
+                        output += f'-{self._int_if_pos(abs(term["num"]))}'
                     else:
-                        output += f' - {Polynomial.int_if_pos(abs(term["num"]))}'
+                        output += f' - {self._int_if_pos(abs(term["num"]))}'
                 else:
                     if i == 0:
                         output += '-'
@@ -254,9 +263,9 @@ class Polynomial:
                 if term['num'] == 1:
                     output += ' + '
                 else:
-                    output += f' + {Polynomial.int_if_pos(term["num"])}'
+                    output += f' + {self._int_if_pos(term["num"])}'
             elif term['num'] != 1:
-                output += str(Polynomial.int_if_pos(term['num']))
+                output += str(self._int_if_pos(term['num']))
 
             i += 1
                 
@@ -265,12 +274,12 @@ class Polynomial:
                     if exponent == 1:
                         output += str(variable)
                     else:
-                        output += variable + Polynomial.super(Polynomial.int_if_pos(exponent))
+                        output += variable + self._super(self._int_if_pos(exponent))
                 
         return output
 
     def __repr__(self):
-        return f'KlebLib.maths.polynomial.Polynomial({self.polynomial})'
+        return f'polynomial.Polynomial({self.polynomial})'
 
     def __float__(self):
         temp = self.polynomial[0].copy()
@@ -313,8 +322,7 @@ class Polynomial:
 
         return equal
 
-    @staticmethod
-    def int_if_pos(num:int|float|str) -> int|float|str:
+    def _int_if_pos(self, num:Union[float, str]) -> Union[int, float, str]:
         #print(f'checking if {num} can be int') #debug
         try:
             assert int(num) == num
@@ -325,10 +333,9 @@ class Polynomial:
         else:
             return int(num)
 
-    @staticmethod
-    def trim_num(string:str, side:str) -> float:
+    def _trim_num(self, string:str, side:str) -> float:
         #print(f'trimming {string}') #debug
-        numPos = Polynomial.get_num_pos(string, side)
+        numPos = self._get_num_pos(string, side)
 
         if numPos:
             #print(f'num found between positions {numPos.start()} and {numPos.end()}') #debug
@@ -337,8 +344,7 @@ class Polynomial:
             #print(f'num not found') #debug
             return 0.0
 
-    @staticmethod
-    def get_num_pos(string:str, side:str) -> list:
+    def _get_num_pos(self, string:str, side:str) -> Union[list, None]:
         if side == 'left':
             numPos = re.search(r'^-?\d+\.?\d*', string)
         else:
@@ -349,8 +355,7 @@ class Polynomial:
         else:
             return None
 
-    @staticmethod
-    def locate(polynomial:list, searchTerm:dict) -> int:
+    def _locate(self, polynomial:list, searchTerm:dict) -> Union[int, bool]:
         #print(f'locating term {searchTerm} in polynomial {polynomial}') #debug
         del searchTerm['num']
         for i, term in enumerate(polynomial):
@@ -360,10 +365,9 @@ class Polynomial:
                 #print(f'found term at position {i}') #debug
                 return i
         #print('didn\'t find term') #debug
-        return None
+        return False
 
-    @staticmethod
-    def consolidate(polynomial:list) -> list:
+    def _consolidate(self, polynomial:list) -> list:
         output = polynomial.copy()
         termsToRemove = []
         
@@ -386,23 +390,21 @@ class Polynomial:
 
         return output
 
-    @staticmethod
-    def super(num:str) -> str:
+    def _super(self, num:str) -> str:
         superscriptMap = {'1': '¹', '2': '²', '3': '³', '4': '⁴', '5': '⁵', '6': '⁶', '7': '⁷', '8': '⁸', '9': '⁹', '-': '⁻'}
 
-        return Polynomial.translate(str(num), superscriptMap)
+        return self._translate(str(num), superscriptMap)
 
-    @staticmethod
-    def translate_super(string:str) -> str:
+    def _translate_super(self, string:str) -> str:
         #print(f'translating {string}') #debug
         superscriptMap = {'1': '¹', '2': '²', '3': '³', '4': '⁴', '5': '⁵', '6': '⁶', '7': '⁷', '8': '⁸', '9': '⁹', '-': '⁻'}
 
         superscriptMap = {v:k for (k, v) in superscriptMap.items()}
 
-        string = Polynomial.translate(string, superscriptMap)
+        string = self._translate(string, superscriptMap)
 
-        num = Polynomial.trim_num(string, 'right')
-        numPos = Polynomial.get_num_pos(string, 'right')
+        num = self._trim_num(string, 'right')
+        numPos = self._get_num_pos(string, 'right')
 
         if numPos:
             if numPos[0] == 0:
@@ -413,30 +415,11 @@ class Polynomial:
                 return string[:numPos[0]] + '^' + str(num)
         else:
             return string
-
-    @staticmethod
-    def translate(string:str, table:dict) -> str:
+        
+    def _translate(self, string:str, table:dict) -> str:
         trans = str.maketrans(
             ''.join(table.keys()),
             ''.join(table.values())
         )
 
         return string.translate(trans)
-
-    @staticmethod
-    def get_variables(polynomial:str|list) -> list:
-        variables = set()
-
-        if type(polynomial) is str:
-            for character in polynomial:
-                # If it is not a number
-                if character not in [str(i) for i in range(10)]:
-                    variables.add(character)
-
-        elif type(polynomial) is list:
-            for term in polynomial:
-                for variable in term:
-                    if variable != 'num':
-                        variables.add(variable)
-
-        return list(variables)
