--- KlebLib/structures/__init__.py
@@ -1,1 +1,0 @@
-from . import series, tree
