+++ KlebLib/universaladdition.py
@@ -1,23 +1,13 @@
-"""uadd -- adds several numbers in a specified base
-classic_add -- adds several numbers in a specified base (deprecated)
-"""
-from .baseconversion import convert_base
-from warnings import warn
+from KlebLib.baseconversion import convert_base
 
-__all__ = ['uadd', 'classic_add']
+__all__ = ['add']
 
-def uadd(base:int|str, *nums:str) -> str:
-    """Add several numbers together in a given base.
-
-    Arguments:
-    base -- the base of the numbers
-    nums -- the numbers to be added
-    """
-    if type(base) is not int and type(base) is not str:
+def uadd(base:int, *nums:str) -> str:
+    if type(base) is not int:
         if type(base) is float:
             base = int(base)
         else:
-            raise ValueError(f'Expected int or str, got {type(base).__name__}')
+            raise ValueError(f'Expected int, got {type(base).__name__}')
 
     nums = list(nums)
     if len(nums) < 2:
@@ -28,19 +18,11 @@ def uadd(base:int|str, *nums:str) -> str:
         del nums[1]
 
     return convert_base(
-        float(convert_base(nums[0], base, 10)) + float(convert_base(nums[1], base, 10)),
+        convert_base(nums[0], base, 10) + convert_base(nums[1], base, 10),
         10, base
     )
 
 def classic_add(base:int, *nums:str) -> str:
-    """Add several numbers together in a given base. Deprecated, do not use. 
-
-    Arguments:
-    base -- the base of the numbers
-    nums -- the numbers to be added
-    """
-    warn('classic_add is deprecated. use uadd instead', DeprecationWarning)
-
     #Check validity of input
     if type(base) is not int:
         if type(base) is float:
@@ -76,8 +58,11 @@ def classic_add(base:int, *nums:str) -> str:
     num2arr = []
   
     #Turn the strings into lists
-    num1 = [i for i in num1]
-    num2 = [i for i in num2]
+    for (num1digit, num2digit) in zip(num1, num2):
+        num1arr.append(num1digit)
+        num2arr.append(num2digit)
+    num1 = num1arr
+    num2 = num2arr
   
     #Reverse the arrays and equalise the lengths
     num1.reverse()
@@ -98,15 +83,15 @@ def classic_add(base:int, *nums:str) -> str:
 		
         result = num1digit + num2digit + carry
         if result >= carry:
-            carry = result // base
-            result = int(result % base)
+        	carry = result // base
+        	result = int(result % base)
         answer.append(result)
         #print(f'result: {result}') #debug
         #print(f'carry: {carry}') #debug
     
     #Handle overflow
     if carry > 0:
-        answer.append(carry)
+    	answer.append(carry)
     
     #Reverse answer to be in the correct order
     #print(f'reverse list answer: {answer}') #debug
@@ -125,6 +110,6 @@ def classic_add(base:int, *nums:str) -> str:
 	
     #Combine digits into one string
     for i in convertedAns:
-        finalAns += str(i)
+    	finalAns += str(i)
     
     return finalAns
