+++ setup.py
@@ -1,11 +1,11 @@
-from setuptools import setup, find_packages
+from setuptools import setup
 
 with open('README.md', 'r') as f:
     long_description = f.read()
 
 setup(
     name='KlebLib',
-    version='1.1.0',
+    version='1.0.2',
     author='Caleb Robson',
     author_email='robson.caleb.299@gmail.com',
     packages=['KlebLib'],
@@ -14,6 +14,5 @@ setup(
     description='A collection of classes and functions that no-one will ever need.',
     long_description=long_description,
     long_description_content_type='text/markdown',
-    python_requires='>=3.10, <4',
-    packages=find_packages(include=['KlebLib', 'KlebLib.*'])
+    python_requires='>=3.8, <4'
 )
